feat: Implement POST endpoint to store short URL

+++ database/migrations/2025_07_15_102102_create_access_counts_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
 
             $table->foreignId('short_url_id')->constrained()->cascadeOnDelete();
-            $table->integer('access_count')->default(0);
+            $table->unsignedBigInteger('access_count')->default(0);
 
             $table->timestamps();
         });
